Remove redundant u16 cast in compute_scroll_offset call site
compute_scroll_offset already returns u16, so the `as u16` at the
render_chat call site was a no-op that clippy's unnecessary_cast lint
(-D warnings) rejects. Left over from the render_chat decomposition.

+++ src/tui/render.rs
@@ -460,7 +460,7 @@ fn render_chat(f: &mut Frame, app: &App, area: Rect) {
                 .border_style(Style::default().fg(Color::Cyan)),
         )
         .wrap(Wrap { trim: false })
-        .scroll((actual_scroll_offset as u16, 0));
+        .scroll((actual_scroll_offset, 0));
 
     f.render_widget(chat, area);
 }
